+++ routes/instock-routes.js
@@ -1,2 +1,10 @@
-const express = require("express");
-const app = express();
+const { Router } = require("express");
+const router = Router();
+
+const postWarehouse = (req, res) => {
+  res.send("hello");
+};
+
+router.post("/warehouse", postWarehouse);
+
+module.exports = router;
