+++ server.js
@@ -2,7 +2,7 @@ const express = require("express");
 const app = express();
 const cors = require("cors");
 const port = 8080;
-const warehouseRoutes = require("./routes/warehouse-routes");
+const warehouseRoutes = require("./routes/instock-routes");
 
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
@@ -12,7 +12,7 @@ app.use(cors());
 app.get("/", function(req, res) {
   res.send("hello world");
 });
-app.post("/warehouse", warehouseRoutes);
+app.use("/warehouse", warehouseRoutes);
 app.listen(port, function() {
   console.log(`app listening on port ${port}`);
 });
